+++ test/plug_rest/conn_test.exs
@@ -4,40 +4,6 @@ defmodule PlugRest.ConnTest do
 
   import PlugRest.Conn
 
-  test "REST body" do
-    resp_body = "Body"
-    conn = conn(:get, "/test")
-    |> put_rest_body(resp_body)
-
-    assert get_rest_body(conn) == resp_body
-  end
-
-  test "get unset REST body" do
-    conn = conn(:get, "/test")
-
-    assert get_rest_body(conn) == nil
-  end
-
-  test "get unset REST body with default" do
-    conn = conn(:get, "/test")
-
-    assert get_rest_body(conn, default: "Default") == "Default"
-  end
-
-  test "media type" do
-    media_type = {"text", "html", %{}}
-    conn = conn(:get, "/test")
-    |> put_media_type(media_type)
-
-    assert get_media_type(conn) == media_type
-  end
-
-  test "get non-existent media type" do
-    conn = conn(:get, "/test")
-
-    assert get_media_type(conn) == ""
-  end
-
   test "parse content type header" do
     content_type = "application/json"